requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
Pillow==10.1.0
opencv-python>=4.8.0
numpy>=1.21.0,<1.25.0
rarfile>=4.0
py7zr>=0.20.0
python-multipart>=0.0.6
paddleocr>=2.7.0
paddlepaddle>=2.5.0